fix oltp_workload
