+++ ydb/tests/datashard/lib/ya.make
@@ -4,8 +4,4 @@ PY_SRCS(
     create_table.py
 )
 
-PEERDIR(
-    ydb/tests/stress/oltp_workload/workload
-)
-
 END()
